# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Listar categorias",
                "description": "Retorna uma lista de categorias com paginação e filtros opcionais.",
                "operationId": "13f3d73cc6bc2b49b3ffb9f4812576e7",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Quantidade de itens por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar categorias pelo nome.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Eletrônicos"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtrar por status ativo (true) ou inativo (false).",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de categorias retornada com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Criar uma nova categoria",
                "description": "Armazena uma nova categoria no banco de dados.",
                "operationId": "7188fdcfb5f38c1d11289a8176a6b558",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "active"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome da categoria.",
                                        "type": "string",
                                        "example": "Eletrônicos"
                                    },
                                    "active": {
                                        "description": "Status ativo ou inativo da categoria.",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Exibir uma categoria",
                "description": "Retorna os detalhes de uma categoria pelo ID.",
                "operationId": "e4710fe1b1bdea207fcf2ee451fb44c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria a ser exibida.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da categoria retornados com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Categoria não encontrada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Atualizar uma categoria",
                "description": "Atualiza os dados de uma categoria existente.",
                "operationId": "e383bdb12623dd341af76d873a632ca6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria a ser atualizada.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome da categoria.",
                                        "type": "string",
                                        "example": "Eletrônicos"
                                    },
                                    "active": {
                                        "description": "Status ativo ou inativo da categoria.",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Categoria atualizada com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Categoria não encontrada."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Deletar uma categoria",
                "description": "Remove uma categoria pelo ID.",
                "operationId": "9bd0148f2ed6d14e14d9dbee9ebb8bb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria a ser deletada.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria deletada com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}